@@ -6,7 +6,7 @@ import com.myprojects.myportfolio.core.dao.BaseDao;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.io.Serial;
 import java.util.HashSet;
 import java.util.Set;
