@@ -2,11 +2,11 @@ package com.myprojects.myportfolio.apigw;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
-@EnableEurekaClient
+@EnableDiscoveryClient
 @ComponentScan(basePackages = {"com.myprojects.myportfolio"})
 public class ApiGWApplication {
 
