@@ -6,7 +6,7 @@ import com.myprojects.myportfolio.core.services.BaseService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
-import javax.transaction.Transactional;
+import org.springframework.transaction.annotation.Transactional;
 
 @Slf4j
 @Service(value = "SkillCategoryService")
